# Fixes 3930: speed up slow rpm searching