--- a/db/migrations/20240425110328_add_trgm_to_rpms.down.sql
+++ b/db/migrations/20240425110328_add_trgm_to_rpms.down.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+DROP EXTENSION IF EXISTS pg_trgm;
+DROP INDEX IF EXISTS rpms_trgm_gin;
+
+COMMIT;