--- a/db/migrations/20240425110328_add_trgm_to_rpms.up.sql
+++ b/db/migrations/20240425110328_add_trgm_to_rpms.up.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+CREATE EXTENSION IF NOT EXISTS pg_trgm;
+CREATE INDEX IF NOT EXISTS rpms_trgm_gin ON rpms USING gin (name gin_trgm_ops);
+
+COMMIT;