@mayurilahane 2024-04-30
lgtm !
deployment is successfully done!

`2024-04-30 08:18:16 [    INFO] [          MainThread] successfully deployed to namespace ephemeral-derewx
2024-04-30 08:18:16 [    INFO] [          MainThread] successful deployment
`

and
I see the mentioned pod running in it 

`content-sources-backend-task-worker-6d85c65c46-gm2g7             1/1     Running     0             13m
`