@mayurilahane 2024-04-30
Logs of the same pod also confirm that migration is done successfully.

```
(new_env) [mlahane@ibm-p8-kvm-01-fsp iqe-content-sources-plugin]$ oc logs content-sources-backend-task-worker-6d85c65c46-gm2g7 -c db-migrate-init 
{"level":"warn","time":"2024-04-30T12:13:38Z","message":"config.yaml file not loaded: Config File \"config.yaml\" Not Found in \"[/configs]\""}
{"level":"debug","platform.content-sources.template":"platform.content-sources.template","time":"2024-04-30T12:13:38Z","message":"internalToReal"}
{"level":"debug","platform.content-sources.template":"platform.content-sources.template","time":"2024-04-30T12:13:38Z","message":"realToInternal"}
{"level":"debug","platform.notifications.ingress":"platform.notifications.ingress","time":"2024-04-30T12:13:38Z","message":"internalToReal"}
{"level":"debug","platform.notifications.ingress":"platform.notifications.ingress","time":"2024-04-30T12:13:38Z","message":"realToInternal"}
{"level":"error","error":"no RDS available","time":"2024-04-30T12:13:38Z","message":"Cannot read RDS CA cert"}
{"level":"warn","time":"2024-04-30T12:13:38Z","message":"Bucket name: content-sources-central-pulp-s3"}
0xc000396570
{"level":"debug","time":"2024-04-30T12:13:39Z","message":"Successfully migrated up"}

```